Basic guard and appoitments loading

+++ backend/src/main/java/com/baranskagata/optometry/entity/Work.java
@@ -5,6 +5,7 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
+import java.time.Duration;
 import java.util.List;
 
 @Entity
@@ -16,7 +17,7 @@ public class Work {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private int id;
+    private Long id;
 
     @Basic
     private String name;
@@ -27,8 +28,10 @@ public class Work {
     @Basic
     private double price;
 
+
     @ManyToMany
     @JoinTable(name = "works_providers", joinColumns = @JoinColumn(name = "id_work"), inverseJoinColumns = @JoinColumn(name = "id_optometrist"))
     private List<Optometrist> providers;
+
 
 }
